feat!: remove legacy v1 server, Code Mode is the only MCP
Drops the tool-per-endpoint v1 server (src/index.ts, helper.ts, logger.ts).
Promotes src/v2/* to src/. lexware-office bin now starts Code Mode;
lexware-office-v2 kept as deprecated alias. Docs rewritten v2-only
(docs/version-guide.md -> docs/guide.md, incl. 1.x migration notes).

BREAKING CHANGE: legacy v1 endpoint tools removed. The lexware-office
binary now runs the Code Mode server (search/execute, read-only by
default). Pin #semver:^1 to keep the legacy server.
