+++ src/docs-contract.test.ts
@@ -5,7 +5,7 @@ import { join, dirname } from 'node:path';
 import test from 'node:test';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
-const versionGuide = readFileSync(join(__dirname, '../../docs/version-guide.md'), 'utf8');
+const guide = readFileSync(join(__dirname, '../docs/guide.md'), 'utf8');
 
 // Split into lines so we can check field names only in voucherlist contexts.
 // A "voucherlist context" is any code block that references /v1/voucherlist.
@@ -13,7 +13,7 @@ const voucherlistCodeBlocks = (() => {
 	const blocks: string[] = [];
 	const fenceRe = /```[\s\S]*?```/g;
 	let match: RegExpExecArray | null;
-	while ((match = fenceRe.exec(versionGuide)) !== null) {
+	while ((match = fenceRe.exec(guide)) !== null) {
 		if (match[0].includes('/v1/voucherlist') || match[0].includes('voucherlist')) {
 			blocks.push(match[0]);
 		}
@@ -21,27 +21,27 @@ const voucherlistCodeBlocks = (() => {
 	return blocks.join('\n');
 })();
 
-test('version-guide voucherlist examples do not reference stale totalGrossAmount field', () => {
+test('guide voucherlist examples do not reference stale totalGrossAmount field', () => {
 	assert.ok(voucherlistCodeBlocks.length > 0, 'Expected at least one voucherlist code block in docs');
 	assert.equal(
 		voucherlistCodeBlocks.includes('totalGrossAmount'),
 		false,
-		'version-guide.md voucherlist example uses stale field totalGrossAmount; use totalAmount instead',
+		'guide.md voucherlist example uses stale field totalGrossAmount; use totalAmount instead',
 	);
 });
 
-test('version-guide voucherlist examples do not reference stale totalNetAmount field', () => {
+test('guide voucherlist examples do not reference stale totalNetAmount field', () => {
 	assert.equal(
 		voucherlistCodeBlocks.includes('totalNetAmount'),
 		false,
-		'version-guide.md voucherlist example uses stale field totalNetAmount; use totalAmount instead',
+		'guide.md voucherlist example uses stale field totalNetAmount; use totalAmount instead',
 	);
 });
 
-test('version-guide voucherlist examples use totalAmount', () => {
+test('guide voucherlist examples use totalAmount', () => {
 	assert.match(
 		voucherlistCodeBlocks,
 		/totalAmount/,
-		'version-guide.md voucherlist example should reference totalAmount',
+		'guide.md voucherlist example should reference totalAmount',
 	);
 });
