+++ src/lexware-spec.ts
