+++ src/truncate.ts
